% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourierin_test.R
\name{dummy_fnc}
\alias{dummy_fnc}
\title{Dummy function to describe arguments}
\usage{
dummy_fnc(f, lower_int, upper_int, lower_eval, upper_eval, const_adj, freq_adj,
  resolution, eval_grid, use_fft)
}
\arguments{
\item{f}{function or a vector of size m. If a function is
provided, it must be able to be evaluated at vectors. If a
vector of values is provided, such evaluations must have been
obtained on a regular grid and the Fourier integral is faster
is m is a power of 2.}

\item{lower_int}{Lower integration limit(s).}

\item{upper_int}{Upper integration limit(s).}

\item{lower_eval}{Lower evaluation limit(s). It can be NULL if an
evaluation grid is provided.}

\item{upper_eval}{Upper evaluation limit(s). It can be NULL if an
evaluation grid is provided.}

\item{const_adj}{Factor related to adjust definition of Fourier
transform. It is usually equal to 0, -1 or 1.}

\item{freq_adj}{Constant to adjust the exponent on the definition
of the Fourier transform. It is usually equal to 1, -1, 2pi or
-2pi.}

\item{resolution}{A vector of integers (faster if powers of two)
determining the resolution of the evaluation grid. Not required
if f is a vector.}

\item{eval_grid}{Optional matrix with d columns with the points
where the Fourier integral will be evaluated. If it is
provided, the FFT will not be used.}

\item{use_fft}{Logical value specifying whether the FFT will be
used.}
}
\description{
The arguments described here will be recycled in the functions of
this package.
}
\author{
Guillermo Basulto-Elias
}
