% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecf_mod}
\alias{ecf_mod}
\title{Module of the empirical characteristic function}
\usage{
ecf_mod(t, smp)
}
\description{
It computes the module of the empirical characteristic function
evaluated on a sample
}

