% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{Times two}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{Vector to be multiplied by two.}
}
\description{
It mulpliplies a vector by two. This is only a test function.
}

