% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fourierin_2d}
\alias{fourierin_2d}
\title{Computes Fourier integral of univariate functions}
\usage{
fourierin_2d(f, a, b, c, d, r, s)
}
\arguments{
\item{f}{Values of the function.}

\item{a}{\eqn{n x 1} vector. Lower integration limit.}

\item{b}{\eqn{n x 1} vector. Upper integration limit.}

\item{c}{\eqn{n x 1} vector. Lower limit of w.}

\item{d}{\eqn{n x 1} vector. Upper limit of w.}

\item{r}{Power in (4.1).}

\item{s}{Scale constant in (4.1).}

\item{m}{Resolution of the integral.}
}
\value{
Continuous Fourier transform values at w.
}
\description{
This function computes univariate and bivariate continuous
Fourier tranform based on the paper by Inverarity (2002):
"Fast computation of multidimensional Fourier integrals".
It is the formula (4.1) on the paper.
}

